refac: docker-compose, dockerfile for db scheme update and user folder refactoring
