+++ .vale.ini
@@ -1,0 +1,9 @@
+StylesPath = .vale-styles
+
+MinAlertLevel = suggestion
+
+[*.md]
+BasedOnStyles = Vale
+
+[*.ts]
+BasedOnStyles = Vale
