+++ apps/auth/src/application/dto/results/object/access-token.object.ts
