+++ apps/auth/src/application/dto/results/object/delete-user.object.ts
