+++ apps/auth/src/application/dto/results/object/token-pair.object.ts
