--- apps/auth/src/application/queries/login-user.query.ts
@@ -1,8 +1,0 @@
-import { IQuery } from '@nestjs/cqrs';
-
-export class LoginUserQuery implements IQuery {
-  constructor(
-    public readonly email: string,
-    public readonly password: string,
-  ) {}
-}
