+++ apps/user/src/application/dto/results/update-user.result.ts
@@ -1,5 +1,5 @@
 import { ObjectType, PickType } from '@nestjs/graphql';
-import { CreateUserResult } from '@user/application/results/create-user.result';
+import { CreateUserResult } from '@user/application/dto/results/create-user.result';
 
 @ObjectType()
 export class UpdateUserResult extends PickType(CreateUserResult, [
